Added in Boss and Enemies
I shifted where the player and camera spawns, to be used in the demo. I
added in a couple of enemies, and the boss at the top level of the
tower.
Still need to add damage model for players and enemies, and probably add
a few more enemies and light sources.
Also need to merge HUD branch.